pyproject.toml
[project]
name = "product-hunt-mailer"
version = "1.0.0"
description = "Daily Product Hunt digest emailer with AI summaries"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "google-generativeai>=0.8.0",
    "resend>=2.0.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
ph-mailer = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
